chore: move execute-provider-utils and get-provider-plugins
