+++ packages/amplify-cli-core/src/extensions/index.ts
@@ -3,7 +3,9 @@ export * from './confirm-prompt';
 export * from './constants';
 export * from './construct-exeInfo';
 export * from './copy-batch';
+export * from './execute-provider-utils';
 export * from './get-all-category-pluginInfos';
+export * from './get-provider-plugins';
 export * from './get-all-envs';
 export * from './get-amplify-appId';
 export * from './get-category-pluginInfo';
