+++ packages/amplify-cli/src/config-steps/c2-configProviders.ts
@@ -1,6 +1,6 @@
 import inquirer, { CheckboxQuestion } from 'inquirer';
 import sequential from 'promise-sequential';
-import { getProviderPlugins } from '../extensions/amplify-helpers/get-provider-plugins';
+import { getProviderPlugins } from 'amplify-cli-core';
 import { normalizeProviderName } from '../input-params-manager';
 
 export async function configProviders(context) {
