+++ packages/amplify-cli/src/extensions/amplify-helpers/remove-env-from-cloud.ts
@@ -2,7 +2,7 @@ import { $TSAny, AmplifyFault } from 'amplify-cli-core';
 import { printer } from 'amplify-prompts';
 import { getProjectConfig } from 'amplify-cli-core';
 import { getAllCategoryPluginInfo } from 'amplify-cli-core';
-import { getProviderPlugins } from './get-provider-plugins';
+import { getProviderPlugins } from 'amplify-cli-core';
 import { raiseInternalOnlyPostEnvRemoveEvent } from '../../execution-manager';
 
 /**
