+++ packages/amplify-cli/src/init-steps/s2-initProviders.ts
@@ -1,7 +1,7 @@
 import { $TSAny, AmplifyError } from 'amplify-cli-core';
 import * as inquirer from 'inquirer';
 import sequential from 'promise-sequential';
-import { getProviderPlugins } from '../extensions/amplify-helpers/get-provider-plugins';
+import { getProviderPlugins } from 'amplify-cli-core';
 import { normalizeProviderName } from '../input-params-manager';
 
 /**
